# Conflicts:
#	build.gradle
#	src/main/java/az/gdg/msarticle/client/AuthenticationClient.java
#	src/main/java/az/gdg/msarticle/config/FeignConfig.java
#	src/main/java/az/gdg/msarticle/controller/ArticleController.java
#	src/main/java/az/gdg/msarticle/exception/WrongDataException.java
#	src/main/java/az/gdg/msarticle/filter/AuthenticationTokenFilter.java
#	src/main/java/az/gdg/msarticle/mapper/ArticleMapper.java
#	src/main/java/az/gdg/msarticle/model/ArticleRequest.java
#	src/main/java/az/gdg/msarticle/model/CommentRequest.java
#	src/main/java/az/gdg/msarticle/model/TagRequest.java
#	src/main/java/az/gdg/msarticle/model/TypeEnum.java
#	src/main/java/az/gdg/msarticle/model/client/auth/UserInfo.java
#	src/main/java/az/gdg/msarticle/model/dto/ArticleDTO.java
#	src/main/java/az/gdg/msarticle/model/dto/CommentDTO.java
#	src/main/java/az/gdg/msarticle/model/dto/TagDTO.java
#	src/main/java/az/gdg/msarticle/model/entity/ArticleEntity.java
#	src/main/java/az/gdg/msarticle/model/entity/CommentEntity.java
#	src/main/java/az/gdg/msarticle/model/entity/TagEntity.java
#	src/main/java/az/gdg/msarticle/repository/ArticleRepository.java
#	src/main/java/az/gdg/msarticle/repository/TagRepository.java
#	src/main/java/az/gdg/msarticle/security/UserAuthentication.java
#	src/main/java/az/gdg/msarticle/service/ArticleService.java
#	src/main/java/az/gdg/msarticle/service/impl/ArticleServiceImpl.java
#	src/main/resources/application.yml
#	src/test/groovy/az/gdg/msarticle/service/impl/ArticleServiceImplTest.groovy

+++ src/main/java/az/gdg/msarticle/model/entity/CommentEntity.java
@@ -4,14 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
-
-import java.time.LocalDateTime;
-import java.util.List;
 
 @Data
 @NoArgsConstructor
@@ -22,12 +23,15 @@ public class CommentEntity {
     @Id
     private String id;
     private String text;
-    private Integer userId;
+    private Long userId;
     private boolean isReply;
+
     @DBRef
     private List<CommentEntity> replies;
+
     @CreatedDate
     private LocalDateTime createdAt;
+
     @LastModifiedDate
     private LocalDateTime updatedAt;
 }
